+++ packages/@codotype/ui/src/components/BlueprintEditor/components/relation/RelationForm.vue
@@ -1,91 +1,42 @@
-
-<template>
+ <template>
   <div class="row">
     <div class="col-lg-12">
-
-      <!-- SOURCE MODEL -->
-      <div class="row">
-        <div class="col-lg-4">
-          <div class="form-group text-center">
-            <label class='mb-0'>{{ selectedSchema.label }}</label>
-            <small class="form-text text-muted">Where the relational data is stored</small>
-            <input type="text" class='form-control' disabled :value="selectedSchema.label">
-          </div>
-        </div>
-
-        <!-- RELATION TYPE -->
-        <!-- CLEANUP - abstract RELATION_TYPE_SELECTOR into independent component? -->
-        <div class="col-lg-4">
-
-          <div class="row">
-            <div class="col-lg-12">
-              <div class="form-group mb-2 text-center">
-                <label class='mb-0'>{{ selectedRelationType.label }}</label>
-                <small class="form-text text-muted mb-0">{{ selectedRelationType.description }}</small>
-              </div>
+      <b-tabs pills>
+        <b-tab title="Cardinality">
+          <div class="row mt-2">
+            <div class="col-sm-12">
+              <p class="lead mb-0">Cardinality</p>
+              <p class="small text-muted"><span class="text-success">Cardinality</span> describes the nature of a <strong>Relation</strong> between two <strong>Schemas</strong></p>
+              <hr />
             </div>
           </div>
+          <RelationTypeForm />
+        </b-tab>
 
-          <div class="row">
-            <div class="col-lg-12">
-              <div class="btn-group w-100">
-                <b-button
-                  v-for="relation in relationTypes"
-                  :key="relation.id"
-                  @click="setRelationType(relation.id)"
-                  size="sm"
-                  variant="outline-primary"
-                  :class="relation.id === model.type ? 'active' : ''"
-                >
-                  <img class='relation-thumbnail' :src=" relation.id === model.type ? 'https://res.cloudinary.com/codotype/image/upload/v1551448517/codotype-icons/' + relation.id.toLowerCase() + '_active' + '.png' : 'https://res.cloudinary.com/codotype/image/upload/v1551448517/codotype-icons/' + relation.id.toLowerCase() + '.png'"/>
-                </b-button>
-              </div>
+        <b-tab title="Alias">
+          <div class="row mt-2">
+            <div class="col-sm-12">
+              <p class="lead mb-0">Alias</p>
+              <p class="small text-muted"><span class="text-success">Aliases</span> enable you to change the way <strong>Schemas</strong> refer to eachother <i>through</i> a <strong>Relation</strong>.</p>
+              <hr />
             </div>
           </div>
-        </div>
+          <div class="row">
+            <div class="col-lg-12">
+              <RelationFormAlias :model="model" />
+            </div>
+          </div>
+        </b-tab>
 
-        <!-- RELATED SCHEMA -->
-        <div class="col-lg-4">
-          <div class="form-group text-center">
-            <label class='mb-0'>Related Schema</label>
-            <small class="form-text text-muted">Schema referenced by this relation</small>
-            <select class="form-control" v-model="model.related_schema_id" @change="updateModel()">
-
-              <!-- <option v-if="model.type === 'HAS_ONE'" v-for="s in allSchemas" :key="s.id" :value="s.id">{{s.label}}</option> -->
-              <template v-if="model.type === 'HAS_MANY'">
-                <option v-for="s in allSchemas" :key="s.id" :value="s.id">
-                  {{s.label_plural}}
-                </option>
-              </template>
-
-              <!-- <option v-if="model.type === 'BELONGS_TO'" v-for="s in allSchemas" :key="s.id" :value="s.id">{{s.label}}</option> -->
-              <template v-else>
-                <option v-for="s in allSchemas" :key="s.id" :value="s.id">
-                  {{s.label}}
-                </option>
-              </template>
-
-            </select>
+        <div class="row">
+          <div class="col-lg-12">
+            <hr>
+          </div>
+          <div class="col-lg-12">
+            <RelationBadge :model="model" />
           </div>
         </div>
-
-        <div class="col-lg-12">
-          <hr>
-        </div>
-
-        <div class="col-lg-12">
-          <RelationFormPreview :model="model" />
-        </div>
-
-        <div class="col-lg-12">
-          <hr>
-        </div>
-
-        <div class="col-lg-12">
-          <RelationFormAlias :model="model" />
-        </div>
-
-      </div>
+      </b-tabs>
 
     </div>
   </div>
@@ -96,18 +47,14 @@
 <script>
 import { mapGetters, mapMutations } from 'vuex'
 import RelationFormAlias from './RelationFormAlias'
-import RelationFormPreview from './RelationFormPreview'
-import RELATION_META from '@codotype/types/lib/relation-meta'
+import RelationTypeForm from './RelationTypeForm'
+import RelationBadge from './RelationBadge'
 
 export default {
   components: {
+    RelationBadge,
+    RelationTypeForm,
     RelationFormAlias,
-    RelationFormPreview
-  },
-  data () {
-    return {
-      relationTypes: Object.keys(RELATION_META).map(rt => RELATION_META[rt])
-    }
   },
   created () {
     // TODO - this should be moved into the relationModule vuex store
@@ -115,45 +62,12 @@ export default {
     const relatedSchema = this.allSchemas.find(m => m.id !== this.selectedSchema.id)
     this.model.schema_id = this.selectedSchema.id
     this.model.related_schema_id = relatedSchema ? relatedSchema.id : this.selectedSchema.id
-    this.updateModel()
+    this.$store.commit('editor/schema/relation/form/model', this.model)
   },
-  methods: {
-    // CLEANUP - this should be moved into the Vuex store
-    setRelationType (relationId) {
-      this.model.type = relationId
-      if (relationId !== 'BELONGS_TO') { this.model.reverse_as = '' }
-      this.updateModel()
-    },
-    updateModel () {
-      this.$store.commit('editor/schema/relation/form/model', this.model)
-    }
-  },
-  computed: {
-    ...mapGetters({
-      model: 'editor/schema/relation/form/model',
-      allSchemas: 'editor/schema/collection/items',
-      selectedSchema: 'editor/schema/selectedModel'
-    }),
-    selectedRelatedSchema () {
-      return this.allSchemas.find(m => m.id === this.model.related_schema_id)
-    },
-    selectedRelationType () {
-      return RELATION_META[this.model.type]
-    }
-  }
+  computed: mapGetters({
+    model: 'editor/schema/relation/form/model',
+    allSchemas: 'editor/schema/collection/items',
+    selectedSchema: 'editor/schema/selectedModel'
+  })
 }
 </script>
-
-<style scoped>
-  img.relation-thumbnail {
-    width: 50%;
-  }
-
-  .btn-outline-primary {
-    width: 33%;
-  }
-
-  .form-text {
-    margin-bottom: 0.6rem;
-  }
-</style>