+++ packages/@codotype/ui/src/components/BlueprintEditor/components/schema/SchemaNewModal.vue
@@ -5,14 +5,14 @@
     :visible="showingModal"
     :title="'New Schema'"
     ok-title='Create Schema'
-    ok-variant='success'
+    ok-variant='primary'
     cancel-title='Cancel'
     cancel-variant='light'
     :ok-disabled="!enableSubmit"
     @ok="createModel()"
     @hide="showModal(false)"
   >
-    <SchemaForm />
+    <SchemaForm :enableSubmit="enableSubmit" :onKeypressEnter="onKeypressEnter" />
   </b-modal>
 </template>
 
@@ -30,6 +30,10 @@ export default {
     showingModal: 'editor/schema/modals/new/showing'
   }),
   methods: {
+    onKeypressEnter() {
+      this.createModel()
+      this.showModal(false)
+    },
     ...mapActions({
       createModel: 'editor/schema/createModel',
     }),